Change signal flow. Add distortion.
The signal flow is now vibrato->distortion->lowpass. This results in a quite
decent sounding one octave, but the rest is mostly crap. It requires further
tuning of the lowpass filter, or even integrating the lowpass with
distortion. Also, the signal flow needs to be rearranged further: the pedals
and percussion aren't supposed to be routed to vibrato.