Many enhancements to the Tonewheel Organ module.
Added configuration settings for vibrato and some of the percussion
parameters (on/off and 2nd/3rd harmonic).  Added a lowpass filter before
distortion to avoid aliasing and muddiness.  Refactored the one-pole lowpass
code to separate file.  Decreased crosstalk, as it previously was a bit
exaggerated.  Removed some misguided dead code.