Fix a number of corner cases in stream player.
The old version used to cause underrun when loop start was near the end of
the block boundary.  The loop point was OK, but the next block was not
requested until it was already too late. Also, loop start buffer is now
preferred to start buffer, to reduce buffer switching.