Drum pattern editor enhancements.
Current velocity is now displayed on the canvas. Notes are displayed as
diamonds. The cursor is now gray and doesn't obscure the notes. Current
edited note property is reset when the note is deleted. Dragging the
cursor out of the window doesn't reset the cursor when a note is being
adjusted.