Fix two seeking-related bugs in sequencer.
First bug: the first time a pattern was played, the events at time zero were
skipped. This was because of the tempo-change logic kicking in and skipping
the zero position (based on incorrect assumption that it has already been played).

Second bug: when switching to a modified pattern, the seek-to-same-place logic
looked for the first event based on PPQN time, which cause the same event
to be played twice.