Fix sampler_layer_clone so that it doesn't clobber the objhdr of regions.
There is still a lot of missing functionality in the set_param area:
it is currently not possible to set params on groups in a way that
would propagate the changes to regions that do not have specific
overrides. This is because the groups don't have a list of children
to update. Also, the changes are still done from the UI thread, including
waveform pointer substitution.

I've added a test to confirm that the UUID clobbering is gone.