Experimental USB MIDI driver-like thingy in Python.
I was mostly interested in the reading part, so writing is very limited
and doesn't handle the F0-FF messages correctly, and sysexes in particular.
Sounds like it's becoming harder and harder to rely on rock-solid operation
of the ALSA/JACK stack, so reinventing the wheel may make some sense.
Also, it's a good opportunity to refresh my memory wrt various details
of the USB protocol.