Implement plug-and-play behaviour for MIDI input devices.
Unfinished implementation - may leak memory. There's lot of debug output
at the moment. It will only work if the USB device in question is accessible
by the current user - either running as root, or having a udev script to
change permissions of the USB device to allow access by the ordinary
user/group. In case of not being able to open or query the device, it will
retry again (and again and again...) so that system-level stuff like udev
can perform its magic.