Improve sync sanity checking in Multimix code.
Previously, the only observed reason for failure of sync-related check was
that the USB connection was unreliable and some iso packets had a non-zero
status. The assertion is now running only if the previous packet has been
received with non-zero length. Also, I've added an assertion for sync packet
size being 3 - just to verify the observations that underlie some assumptions
in the code.