Attempt to separate 'runtime' and 'work' versions of sampler layer data.
The idea is that a sequence of several set_param commands is
followed by a call to sampler_layer_update, which updates the voices
in realtime thread to use the new copy of the data.

This is untested and some parts of it will be changed anyway - for example,
there is no way to remove settings from a region in a way that restores
group-level settings.