Add Python API to add/remove master track items.
A single function is used to add, change and remove master track items at a
specific position. Setting both tempo and time signature to 0 removes an item.
Passing None (or -1.0 for tempo and -1/N for time signature) leaves that part
unchanged.

The 0 values are not yet handled correctly on sequencer side, while building
the tempo map. The code will need to be updated to borrow missing values from
the previous items or from current master setting.