Make sure all voices are released when sampler is destroyed.
This is an old bug, it hasn't been caused by streaming code. However, it is
particularly harmful with streaming, as all unreleased voices keep prefetch
pipes in active state, and trigger an assertion when deleting the prefetch
stack. Besides, the prefetch stack is released before the programs, which will
be fixed in the next commit. This is deliberately kept streaming-free to allow
for cherry-pick into non-streaming branch.