First half-working version of ad-hoc pattern playback.
This is sort of good enough for testing the API, but leaks memory and increases
CPU load every time a pattern is played. I'm planning to fix it very soon, as
it is not really usable in present form.

The ad-hoc patterns have their own tempo and are not in any way synced to song
or JACK, so they may be used for things like chord preview or ornament preview
in Laborejo, or for 'light' playback uses that involve standalone patterns
(MIDI-baed game sound effects, music-oriented educational software etc.)