Implement lazy clean-up for ad-hoc patterns.
The finished ad-hoc pattern structures will be disposed of the next time an
adhoc pattern is played, or in scene destructor. This should be sufficient in
practice, except if a large number of patterns are started at the same time,
and then no patterns are played for the rest of the lifetime of the scene -
in that case, the stopped patterns' outputs will remain in the scene input
merger and unnecessarily consume CPU time.