Implement new, ctypes-based interface for using calfbox with Python.
The new interface eliminates the need for a Python shared library. Instead
it loads the calfbox executable itself as a shared library, and uses new
C API (cbox_embed_*) instead of the old Python extension.

Because of all the translation between Python function calls, the OSC
interface and the C functions that do the actual work, the new interface
is probably not as fast as the old one, but ultimately the OSC interface
may be abandoned in future, in favour of direct calls to C code.

An additional bonus from all this work is that it may be possible to
use other languages instead of Python 3. In fact, the _cbox2 module
can be used with both Python 2 and 3 - however, the high-level wrappers
in cbox still require Python 3 because of their use of the new language
features like metaclasses.