Implement a 'prevoices' mechanism in sampler, use for region delay.
This provides a way for deferring allocating a voice until the region
actually has to output audio, and only keeping a lightweight data
structure for those future notes. So far this is only used to implement
the trivial 'region delay' SFZ feature, but could be a base of all
sorts of mini-sequencer/arpeggiator functionality, and also can be
used to efficiently implement sync_offset/sync_beats in SFZv1.