New data model for SFZ regions + better test. Beware of new bugs!
The new data model uses a consistent representation for the region
parameters of all types. This should make it easier to add new
parameters, and on its own it brings faster parsing and basic data
validation. It also fixes some remaining group-region data
propagation bugs.

At the same, this is an all-new code, so it may well be full of
new bugs. The propagation logic and some of the validation is
tested via a unit test, but the impact on audio synthesis
has only been tested by loading a few SFZ files I had at hand.