Add more SFZ tests. Fix various bug.
A regression was caused by the RLL optimization to create per-note-range
region lists, and caused sw_last keyswitches to not work. This is now fixed
by a somewhat clumsy logic that adds release regions to the note-on lists
just so that the keyswitches are triggered. There are better ways to handle
it, but that's for future.

One bug was caused by a misunderstanding of the spec about the oncc behaviour.
According to the SFZ test kit, the region should trigger on every CC with
the value within the specified range, not just when transitioning from
out-of-range to in-range.

Another bug caused a CC message to not be processed when there was a
region with on_*cc somewhere in the program but not for that particular
controller number, due to a premature return from a function. So, the
existence of on_*cc regions caused other controllers (default or
user-defined) to not work.

The tests added cover the correct oncc triggering behaviour and also
the up/down/last key switches.