New, more CPU-efficient handling of sw_last keyswitch regions + test.
This works best for a single keyswitch group, but can handle up to 16
of them in a single file. More keyswitch groups use more CPU
time and RAM, even for the notes that don't depend on keyswitch groups.
However, in typical case the impact should be pretty low, and the
notes where the switching does occur will be found much quicker,
as the regions that are switched off are off the lookup list.